Graph now builds!
Just need to get it to run. Erroring out on /home/rujak/Documents/Dev/Tensorlfow/OS-P-Test/src/trainer.py
with a ValueError: Cannot use the given session to execute operation: the operation's graph is different from the session's graph. Wanted to push it to make sure I can't lose it.